Extract getListMatcher into src/trusted-process/list-matcher.ts
The PolicyEngine runtime-imported getListMatcher from
src/pipeline/dynamic-list-types.ts, crossing from the security kernel
into the offline pipeline layer. This split the file's two concerns:

- Compile-time: validators, LIST_TYPE_REGISTRY, ListTypeDef -- still
  used by src/pipeline/list-resolver.ts. Stays in dynamic-list-types.ts.
- Runtime: getListMatcher -- only consumer is the policy engine. Moves
  to src/trusted-process/list-matcher.ts, alongside its dependency
  domainMatchesAllowlist.

Pure move -- no behavior change, no signature change. Updates the one
runtime importer (policy-engine.ts) and the test file. Drops the
now-unused domainMatchesAllowlist import from dynamic-list-types.ts and
refreshes the file-header JSDoc to point at the new location.

Layer-violation refactor backlog item #3.
